Updating brotli submodules to build with GCC 11.
GCC 1.1.0 introduced a new warning for variable length arrays as
parameters (vla-parameter). Brotli compression library had issues with
this which were fixed upstream by commits
0a3944c8c99b8d10cc4325f721b7c273d2b41f7b and
27dd7265403d8e8fed99a854b9c3e1db7d79525f in broli repository. Updating
edk2 submodules to have these commits. While at it, adding two more
commits from the brotli repository to make it compatible with future
versions of python as well.

The addition of new commits to the brotli tree made it necessary to add
more files to the build process of the brotli compression library in
BaseTools package.

Signed-off-by: Leonardo Garcia <leonardo.garcia@linaro.org>